Move some thread pool code to sources
